requirements.txt
Django
mysqlclient
django-easy-pdf
git+https://github.com/chrisglass/xhtml2pdf.git